+++ telegram-bot/src/main/java/pro/sky/telegrambot/repository/NotificationTaskRepository.java
@@ -3,5 +3,5 @@ package pro.sky.telegrambot.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pro.sky.telegrambot.model.NotificationTask;
 
-public interface NotificationTaskRepository extends JpaRepository<NotificationTask, Long> {
+public interface NotificationTaskRepository extends JpaRepository<NotificationTask, Integer> {
 }
